# [PROF-10128] Refactor and speed up profiler stack sampling