--- a/ext/datadog_profiling_native_extension/collectors_stack.h
+++ b/ext/datadog_profiling_native_extension/collectors_stack.h
@@ -23,5 +23,5 @@ void record_placeholder_stack(
   ddog_CharSlice placeholder_stack
 );
 uint16_t sampling_buffer_check_max_frames(int max_frames);
-sampling_buffer *sampling_buffer_new(uint16_t max_frames);
+sampling_buffer *sampling_buffer_new(uint16_t max_frames, ddog_prof_Location *locations);
 void sampling_buffer_free(sampling_buffer *buffer);