… ext/datadog_profiling_native_extension/datadog_ruby_common.h
… ext/datadog_profiling_native_extension/extconf.rb
… ext/datadog_profiling_native_extension/heap_recorder.c
… ext/datadog_profiling_native_extension/private_vm_api_access.c
… ext/datadog_profiling_native_extension/private_vm_api_access.h
… ext/libdatadog_api/datadog_ruby_common.h
… spec/datadog/profiling/stack_recorder_spec.rb