**What does this PR do?**

This PR refactors the way the profiler samples the Ruby stack, and then builds on that refactoring to speed up sampling.

It also opens the door to future speedups once libdatadog allows us to skip re-interning strings, frames and stacks that we know have been sampled before.

Specifically this PR does the following:
1. It tweaks the data that `ddtrace_rb_profile_frames` returns from sampling the Ruby stack so that the stack collector can use cheaper APIs to get the same information as before.
2. It changes the sampling buffer we put stack traces in to be per-thread. This allows us to reuse some of the work we do from sample to sample, in cases where frames on the stack are the same. Thus sampling the same stack over and over will be slightly cheaper.
3. It adds a few more micro-optimizations that I spotted from profiling the profiler :)
4. When we sample stacks, we expose a flag that can be used to determine if a frame was unchanged from the last sample. This is what opens the door for future optimizations using libdatadog.

Here's the results of running benchmarks/profiler_sample_loop_v2.rb with these changes on my local machine:

```
Warming up --------------------------------------
stack collector improved-sampling
                       783.000 i/100ms
Calculating -------------------------------------
stack collector improved-sampling
                          7.828k (± 3.3%) i/s -    234.900k in  30.042969s

Warming up --------------------------------------
stack collector master
                       746.000 i/100ms
Calculating -------------------------------------
stack collector master
                          7.057k (± 3.3%) i/s -    211.864k in  30.058848s

Comparison:
stack collector improved-sampling:     7827.8 i/s
stack collector master:     7056.8 i/s - 1.11x  slower
```

**Motivation:**

This PR further reduces the overhead of profiling, especially more complex/deep stacks. I started looking into this with the goal of improving allocation profiling, but stack sampling is shared with cpu/wall-time profiling as well, so those will also be able to take advantage of the improvement.

**Additional Notes:**

There's a lot of small parts moving around in this PR. I recommend reviewing commit-by-commit, as I tried to break every change into a small standalone step that can be reviewed in a much nicer way.

**How to test the change?**

Our existing tests cover these changes.